Create starrer.py
A simple python application created for getting to know the click and request libraries.
The application is for viewing and manipulating with the stars you can give here on Github. Command show will show you if you gave a star or a selected repository, add will add it and remove will delete it. Simple :]. 
A Github token is needed for proper authentication. For this code to work for you, you should save the token into an 'auth.cfg' file with the following syntax:
[github]
token = 1548955fsafafwfefwefwfw156615156efw
You should also save this file in the same directory as you have this python file.
Enjoy!